fix: All gcc warnings

+++ src/main.c
@@ -5,7 +5,8 @@
 
 int main(const int argc, const char* const* const argv) {
     /* Get the port to listen to */
-    unsigned short port = argc > 1 ? atoi(argv[1]) : DEFAULT_PORT;
+    unsigned short port = argc > 1 ? (unsigned short)atoi(argv[1])
+                                   : DEFAULT_PORT;
     if (port == 0) port = DEFAULT_PORT;
 
     /* Try to open the socket connection */
